% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watershed.R
\name{calc_grad}
\alias{calc_grad}
\title{Wraps RCMG calculation}
\usage{
calc_grad(data, type = "euclidean", r = 1)
}
\arguments{
\item{data}{HSI_data or array to calculate gradient of.}

\item{type}{"euclidean" or "cos".}

\item{r}{how many pairs of pixels to remove (between 0 and 4).}
}
\value{
Appropriate RCMG gradient
}
\description{
By default, if data is HSI_data, it will calculate gradient for data$img.
For more control, you can pass in an array of 3 dimensions.
Support Euclidean or Cosine distance for type.
}
\examples{
\dontrun{
grad <- calc_grad(data)
grad_c <- calc_grad(data$img_raw, "cos", 2)
}
}
