% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{get_all_data}
\alias{get_all_data}
\title{Get all data function}
\usage{
get_all_data(...)
}
\arguments{
\item{...}{Passed to get_data}
}
\description{
Downloads HSI data for all 6 images
}
\examples{
\dontrun{
get_all_data(verbose = FALSE)
}

}
