% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{get_data}
\alias{get_data}
\title{Get data function}
\usage{
get_data(name, folder = "./HSI_Data/", verbose = TRUE, clip_p = 0.9975)
}
\arguments{
\item{name}{image name to be downloaded}

\item{folder}{folder to send output to}

\item{verbose}{boolean determining if user wants function messages}

\item{clip_p}{percentile to clip the image to}
}
\value{
HSI_data object
}
\description{
Downloads HSI data, can be indianpines, salinas, paviau (university),
paviac (center), ksc (Kennedy Space Center), botswana
}
\examples{
\dontrun{
bots <- get_data("botswana")
pu <- get_data("PaviaU", verbose = FALSE, clip_p = .9950)
}

}
