% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_neigh}
\alias{get_neigh}
\title{Get neighborhood of a pixel in a 3D image}
\usage{
get_neigh(row, col, img)
}
\arguments{
\item{row}{An integer representing the row index of the pixel of interest.}

\item{col}{An integer representing the column index of the pixel of interest.}

\item{img}{A cube object containing the image data.}
}
\value{
A matrix representing the neighborhood of the pixel of interest.
}
\description{
This function returns the neighborhood of a pixel in a 3D image. Returns the
8 neighboring (plus center pixel) for the pixel at row and col. The input
\code{img} is a three-dimensional array containing the image data.
}
\examples{
img <- array(1:27, dim = c(3, 3, 3))
get_neigh(2, 2, img)

}
