% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{plot.HSI_data}
\alias{plot.HSI_data}
\title{Plot false color image}
\usage{
\method{plot}{HSI_data}(x, y, ...)
}
\arguments{
\item{x}{HSI_data}

\item{y}{not used}

\item{...}{not used}
}
\value{
nothing
}
\description{
Plot false color image
}
