% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{plot.HSI_grad}
\alias{plot.HSI_grad}
\title{Plot image gradient}
\usage{
\method{plot}{HSI_grad}(x, y, ...)
}
\arguments{
\item{x}{HSI_grad object}

\item{y}{not used}

\item{...}{can include log to plot on the log_10 scale}
}
\value{
Nothing
}
\description{
Plot image gradient
}
