% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{plot.HSI_seg}
\alias{plot.HSI_seg}
\title{Plot segmentation}
\usage{
\method{plot}{HSI_seg}(x, y, ...)
}
\arguments{
\item{x}{HSI_seg object}

\item{y}{HSI_data object which was used to create HSI_seg}

\item{...}{not used yet}
}
\value{
nothing
}
\description{
Plot segmentation
}
