% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcmg_euclid}
\alias{rcmg_euclid}
\title{Calculates RCMG using Euclidean distance}
\usage{
rcmg_euclid(img, r = 1L)
}
\arguments{
\item{img}{A 3D cube containing the image data}

\item{r}{An integer specifying the radius of the rolling window}
}
\value{
A matrix containing the Euclidean distances of the pixels
}
\description{
For every pixel, find the RCMG where the 8 neighboring pixels make up the
pairwise distance matrix, and r is the number of pairs of pixels to remove.
Don't set r larger than 8.
}
