% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{relabel_seg}
\alias{relabel_seg}
\title{Relabels segmentation}
\usage{
relabel_seg(seg)
}
\arguments{
\item{seg}{a segmentation matrix to relabel.}
}
\value{
relabeled matrix.
}
\description{
Relabels segmentation map starting the labeling at 1.
}
