% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{summary.HSI_data}
\alias{summary.HSI_data}
\title{Get Summary of the data}
\usage{
\method{summary}{HSI_data}(object, ...)
}
\arguments{
\item{object}{HSI_data}

\item{...}{not used}
}
\value{
nothing
}
\description{
Get Summary of the data
}
