% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watershed.R
\name{watershed_hsi}
\alias{watershed_hsi}
\title{Watershed for HSI}
\usage{
watershed_hsi(grad, tolerance = 0.01, ext = 200, ...)
}
\arguments{
\item{grad}{The gradient of an HSI image, typically returned by one of
the RCMG functions.}

\item{tolerance}{The tolerance to be used by the watershed algorithm.}

\item{ext}{A smoothing parameter to be used in the watershed algorithm}

\item{...}{passed to calc_grad and EBImage::watershed}
}
\value{
A segmented image built from the gradient, in the form of a matrix.
}
\description{
Run the watershed algorithm on the gradient of an HSI image.
}
