% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{get_data}
\alias{get_data}
\title{Get data function}
\usage{
get_data(name, folder = "./HSI_Data/", verbose = TRUE, clip_p = 0.9975)
}
\arguments{
\item{name}{String argument of image name to be downloaded.}

\item{folder}{String argument of folder to send output to.}

\item{verbose}{boolean argument that determines if messages are printed as
function runs.}

\item{clip_p}{Numeric argument between zero and one that represents
percentile to clip the image outliers to.}
}
\value{
HSI_data object
}
\description{
Downloads or loads previously downloaded HSI data. Data can be indianpines,
salinas, paviau (university), paviac (center), ksc (Kennedy Space Center) or
botswana. The function returns an HSI_data object that can be used for
image segmentation, and plotting the resulting segmentation.
}
\examples{
\dontrun{
bots <- get_data("botswana")
pu <- get_data("PaviaU", verbose = FALSE, clip_p = .9950)
}

}
